+++ intl/icu/source/i18n/precision.cpp
@@ -239,7 +239,7 @@ FixedPrecision::initVisibleDigits(
         }
     }
     
-    if (n >= 0 && initVisibleDigits(scaled, -n, digits, status)) {
+    if (n >= 0 && initVisibleDigits(static_cast<int64_t>(scaled), -n, digits, status)) {
         digits.fAbsDoubleValue = fabs(value);
         digits.fAbsDoubleValueSet = U_SUCCESS(status) && !digits.isOverMaxDigits();
